fix: reorganized schema into separate files per table
