+++ src/models/Organization.ts
@@ -1,0 +1,34 @@
+import {
+  bigint,
+  pgTable,
+  text,
+  timestamp,
+  uniqueIndex,
+} from 'drizzle-orm/pg-core';
+
+export const organizationSchema = pgTable(
+  'organization',
+  {
+    id: text('id').primaryKey(),
+    stripeCustomerId: text('stripe_customer_id'),
+    stripeSubscriptionId: text('stripe_subscription_id'),
+    stripeSubscriptionPriceId: text('stripe_subscription_price_id'),
+    stripeSubscriptionStatus: text('stripe_subscription_status'),
+    stripeSubscriptionCurrentPeriodEnd: bigint(
+      'stripe_subscription_current_period_end',
+      { mode: 'number' },
+    ),
+    updatedAt: timestamp('updated_at', { mode: 'date' })
+      .defaultNow()
+      .$onUpdate(() => new Date())
+      .notNull(),
+    createdAt: timestamp('created_at', { mode: 'date' }).defaultNow().notNull(),
+  },
+  (table) => {
+    return {
+      stripeCustomerIdIdx: uniqueIndex('stripe_customer_id_idx').on(
+        table.stripeCustomerId,
+      ),
+    };
+  },
+);
