+++ src/models/Schema.ts
@@ -1,11 +1,5 @@
-import {
-  bigint,
-  pgTable,
-  serial,
-  text,
-  timestamp,
-  uniqueIndex,
-} from 'drizzle-orm/pg-core';
+import { organizationSchema } from './Organization';
+import { todoSchema } from './ToDo';
 
 // This file defines the structure of your database tables using the Drizzle ORM.
 
@@ -17,41 +11,7 @@ import {
 // The migration is automatically applied during the next database interaction,
 // so there's no need to run it manually or restart the Next.js server.
 
-export const organizationSchema = pgTable(
-  'organization',
-  {
-    id: text('id').primaryKey(),
-    stripeCustomerId: text('stripe_customer_id'),
-    stripeSubscriptionId: text('stripe_subscription_id'),
-    stripeSubscriptionPriceId: text('stripe_subscription_price_id'),
-    stripeSubscriptionStatus: text('stripe_subscription_status'),
-    stripeSubscriptionCurrentPeriodEnd: bigint(
-      'stripe_subscription_current_period_end',
-      { mode: 'number' },
-    ),
-    updatedAt: timestamp('updated_at', { mode: 'date' })
-      .defaultNow()
-      .$onUpdate(() => new Date())
-      .notNull(),
-    createdAt: timestamp('created_at', { mode: 'date' }).defaultNow().notNull(),
-  },
-  (table) => {
-    return {
-      stripeCustomerIdIdx: uniqueIndex('stripe_customer_id_idx').on(
-        table.stripeCustomerId,
-      ),
-    };
-  },
-);
-
-export const todoSchema = pgTable('todo', {
-  id: serial('id').primaryKey(),
-  ownerId: text('owner_id').notNull(),
-  title: text('title').notNull(),
-  message: text('message').notNull(),
-  updatedAt: timestamp('updated_at', { mode: 'date' })
-    .defaultNow()
-    .$onUpdate(() => new Date())
-    .notNull(),
-  createdAt: timestamp('created_at', { mode: 'date' }).defaultNow().notNull(),
-});
+export {
+  organizationSchema,
+  todoSchema,
+};
